Include validate cog into cog function add header
